fix warnings (#100)
fix warnings

+++ docs/examples/c/utils/base64.h
@@ -19,7 +19,7 @@
 */
 
 #include <stdint.h>
-
+#include <string.h>
 /**
  * Decode a base64-encoded string.
  *
@@ -41,7 +41,7 @@ int base64_decode(uint8_t *out, const char *in, int out_size);
  * @param in_size  size in bytes of the 'in' buffer
  * @return         'out' or NULL in case of error
  */
-char *base64_encode(char *out, int out_size, const uint8_t *in, int in_size);
+char *base64_encode(char *out, size_t out_size, const uint8_t *in, size_t in_size);
 
 /**
  * Calculate the output size needed to base64-encode x bytes.
